+++ asyncgit/src/sync/cred.rs
@@ -2,6 +2,7 @@
 
 use super::{
 	remotes::{
+		get_default_remote_for_fetch_in_repo,
 		get_default_remote_for_push_in_repo,
 		get_default_remote_in_repo,
 	},
@@ -49,6 +50,26 @@ pub fn need_username_password(repo_path: &RepoPath) -> Result<bool> {
 }
 
 /// know if username and password are needed for this url
+/// TODO: Very similar to `need_username_password_for_fetch`. Can be refactored. See also
+/// `need_username_password`.
+pub fn need_username_password_for_fetch(
+	repo_path: &RepoPath,
+) -> Result<bool> {
+	let repo = repo(repo_path)?;
+	let remote = repo
+		.find_remote(&get_default_remote_for_fetch_in_repo(&repo)?)?;
+	let url = remote
+		.url()
+		.or_else(|| remote.url())
+		.ok_or(Error::UnknownRemote)?
+		.to_owned();
+	let is_http = url.starts_with("http");
+	Ok(is_http)
+}
+
+/// know if username and password are needed for this url
+/// TODO: Very similar to `need_username_password_for_fetch`. Can be refactored. See also
+/// `need_username_password`.
 pub fn need_username_password_for_push(
 	repo_path: &RepoPath,
 ) -> Result<bool> {
@@ -93,6 +114,36 @@ pub fn extract_username_password(
 }
 
 /// extract username and password
+/// TODO: Very similar to `extract_username_password_for_fetch`. Can be refactored.
+pub fn extract_username_password_for_fetch(
+	repo_path: &RepoPath,
+) -> Result<BasicAuthCredential> {
+	let repo = repo(repo_path)?;
+	let url = repo
+		.find_remote(&get_default_remote_for_fetch_in_repo(&repo)?)?
+		.url()
+		.ok_or(Error::UnknownRemote)?
+		.to_owned();
+	let mut helper = CredentialHelper::new(&url);
+
+	//TODO: look at Cred::credential_helper,
+	//if the username is in the url we need to set it here,
+	//I dont think `config` will pick it up
+
+	if let Ok(config) = repo.config() {
+		helper.config(&config);
+	}
+
+	Ok(match helper.execute() {
+		Some((username, password)) => {
+			BasicAuthCredential::new(Some(username), Some(password))
+		}
+		None => extract_cred_from_url(&url),
+	})
+}
+
+/// extract username and password
+/// TODO: Very similar to `extract_username_password_for_fetch`. Can be refactored.
 pub fn extract_username_password_for_push(
 	repo_path: &RepoPath,
 ) -> Result<BasicAuthCredential> {