+++ asyncgit/src/sync/mod.rs
@@ -79,8 +79,9 @@ pub use merge::{
 };
 pub use rebase::rebase_branch;
 pub use remotes::{
-	get_default_remote, get_default_remote_for_push, get_remotes,
-	push::AsyncProgress, tags::PushTagsProgress,
+	get_default_remote, get_default_remote_for_fetch,
+	get_default_remote_for_push, get_remotes, push::AsyncProgress,
+	tags::PushTagsProgress,
 };
 pub(crate) use repository::repo;
 pub use repository::{RepoPath, RepoPathRef};